+++ audio_device/sdl/SdlAudioDevice.hpp
@@ -18,13 +18,13 @@ namespace GBA::audio {
 		void Start() override;
 		void Stop() override;
 
-		void PushSamples(common::i8* samples) override;
+		void PushSamples(common::i16* samples) override;
 
 		friend void feed_callback(void* userdata, uint8_t* stream, int len);
 
 	private :
 		std::mutex m_buffer_mutex;
-		i8* m_buffer;
+		i16* m_buffer;
 
 		SDL_AudioSpec m_spec;
 		SDL_AudioDeviceID m_dev_id;
